Introduce libclaimtrie

+++ src/miner.cpp
@@ -9,6 +9,7 @@
 #include <chain.h>
 #include <chainparams.h>
 #include <claimscriptop.h>
+#include <claimtrie/forks.h>
 #include <coins.h>
 #include <consensus/consensus.h>
 #include <consensus/tx_verify.h>
@@ -204,7 +205,7 @@ std::unique_ptr<CBlockTemplate> BlockAssembler::CreateNewBlock(const CScript& sc
     {
         CClaimTrieCache trieCache(pclaimTrie);
         blockToCache(pblock, trieCache, nHeight);
-        pblock->hashClaimTrie = trieCache.getMerkleHash();
+        pblock->hashClaimTrie = uint256(trieCache.getMerkleHash());
     }
     CValidationState state;
     if (!TestBlockValidity(state, chainparams, *pblock, pindexPrev, false, false)) {
